% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_exiftool}
\alias{configure_exiftool}
\alias{configure_perl}
\alias{configure_exiftool_reset}
\title{Configure perl, ExifTool}
\usage{
configure_exiftool(
  command = NULL,
  perl_path = NULL,
  install_url = NULL,
  install_location = NULL,
  quiet = FALSE
)

configure_perl(perl_path = NULL, quiet = FALSE)

configure_exiftool_reset()
}
\arguments{
\item{command}{The exiftool command or location of exiftool.pl}

\item{perl_path}{The path to the perl executable}

\item{install_url}{The url from which exiftool could be installed}

\item{install_location}{The location to install exiftool}

\item{quiet}{Use quiet = FALSE to display status updates}
}
\value{
The exiftool command, invisibly
}
\description{
Configure perl, ExifTool
}
