% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getexif.R
\name{exiftool.call}
\alias{exiftool.call}
\title{Call exiftool from R}
\usage{
exiftool.call(fnames = NULL, args = c("--help"))
}
\arguments{
\item{fnames}{a list of filenames (\code{shQuote()} will be applied to this vector)}

\item{args}{a list of non-shell quoted arguments (e.g. \code{-n -csv})}
}
\value{
the output generated by exiftool.
}
\description{
Call exiftool from R
}
\examples{
exiftool.call()
}

