% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exif.R
\name{exiftool_call}
\alias{exiftool_call}
\title{Call exiftool from R}
\usage{
exiftool_call(args = c("--help"), fnames = NULL, intern = FALSE, ...,
  quiet = FALSE)
}
\arguments{
\item{args}{a list of non-shell quoted arguments (e.g. \code{-n -csv})}

\item{fnames}{a list of filenames (\code{shQuote()} will be applied to this vector)}

\item{intern}{\code{TRUE} if output should be returned as a character vector.}

\item{...}{additional arguments to be passed to \code{system()}}

\item{quiet}{Suppress output of the command itself.}
}
\value{
The exit code if \code{intern=FALSE}, or the standard output as a character vector
 if \code{intern=TRUE}.
}
\description{
Uses \code{system()} to run a basic call to \code{exiftool}.
}
\examples{
exiftool_call("--help")

}
