% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rational_to_numeric}
\alias{rational_to_numeric}
\title{rational_to_numeric}
\usage{
rational_to_numeric(x)
}
\arguments{
\item{x}{A vector of character strings containing rational numbers to convert.}
}
\value{
A vector of floating point values corresponding to \code{x}.
}
\description{
Converts rational numbers from RATIONAL/character string to a floating point representation
}
\examples{
rational_to_numeric("1/3200")
}
\seealso{
\code{\link{read_exif_tags}}
}
\author{
Charles Martin & Andrew G. Brown
}
