% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rationalDMS_to_decimal}
\alias{rationalDMS_to_decimal}
\title{Convert rational number degrees, minutes, seconds to decimal degrees.}
\usage{
rationalDMS_to_decimal(x, .split = " ")
}
\arguments{
\item{x}{Coordinates strings (rational degrees, minutes and seconds) to convert.}

\item{.split}{Delimiter separating the rational numbers (degrees, minutes
and seconds in \code{x}).}
}
\value{
Coordinates in decimal degrees. A vector of floating point values.
}
\description{
Convert sets of three rational numbers representing degrees,
minutes, and seconds to floating point values reflecting coordinate in
decimal degrees. Takes vector of character strings with each element of \code{x} containing three rational numbers separated by \code{.split}
}
\examples{
rationalDMS_to_decimal(c("120/1 17/1 1571/50"))
}
\seealso{
\code{\link{read_exif_tags}}
}
\author{
Andrew G. Brown
}
