% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exifr.R
\name{read_exif_tags}
\alias{read_exif_tags}
\title{Extract EXIF tags from an image file.}
\usage{
read_exif_tags(file_path)
}
\arguments{
\item{file_path}{The path to the image.}
}
\value{
A list of EXIF tags and their values.
}
\description{
Values are returned directly from the file, without any formatting.
For example, the exposure time (ExposureTime), will be "1/3200".
This value can be converted afterwards with \code{\link{rational_to_numeric}}.
}
\examples{
read_exif_tags(system.file("extdata", "preview.jpg", package = "EXIFr"))[["ExposureTime"]]
}
\seealso{
\code{\link{rational_to_numeric}}
}

