% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exifr.R
\name{supported_tags}
\alias{supported_tags}
\title{List EXIF tags currently supported by this package.}
\usage{
supported_tags()
}
\value{
A vector of EXIF tag names.
}
\description{
Currently supported tags include:
ApertureValue, Artist, BitsPerSample, BrightnessValue, CellLength,
CellWidth, CFAPattern, ColorMap, ColorSpace, ComponentsConfiguration,
CompressedBitsPerPixel, Compression, Contrast, Copyright, CustomRendered,
DateTime, DateTimeDigitized, DateTimeOriginal, DeviceSettingDescription,
DigitalZoomRatio, ExifVersion, ExposureBiasValue, ExposureIndex,
ExposureMode, ExposureProgram, ExposureTime, ExtraSamples, FileSource,
FillOrder, Flash, FlashEnergy, FlashpixVersion, FNumber, FocalLength,
FocalLengthIn35mmFilm, FocalPlaneResolutionUnit, FocalPlaneXResolution,
FocalPlaneYResolution, FreeByteCounts, FreeOffsets, GainControl,
GPSAltitude, GPSAltitudeRef, GPSAreaInformation, GPSDateStamp,
GPSDestBearing, GPSDestBearingRef, GPSDestDistance, GPSDestDistanceRef,
GPSDestLatitude, GPSDestLatitudeRef, GPSDestLongitude, GPSDestLongitudeRef,
GPSDifferential, GPSDOP, GPSImgDirection, GPSImgDirectionRef, GPSLatitude,
GPSLatitudeRef, GPSLongitude, GPSLongitudeRef, GPSMapDatum, GPSMeasureMode,
GPSProcessingMethod, GPSSatellites, GPSSpeed, GPSSpeedRef, GPSStatus,
GPSTimeStamp, GPSTrack, GPSTrackRef, GPSVersionID, GrayResponseCurve,
GrayResponseUnit, HostComputer, ImageDescription, ImageLength,
ImageUniqueID, ImageWidth, ISOSpeedRatings, LightSource, Make,
MakerNote, MaxApertureValue, MaxSampleValue, MeteringMode, MinSampleValue,
Model, NewSubfileType, OECF, OffsetTime, OffsetTimeDigitized,
OffsetTimeOriginal,  Orientation, PhotometricInterpretation,
PixelXDimension, PixelYDimension, PlanarConfiguration, RelatedSoundFile,
ResolutionUnit, RowsPerStrip, SamplesPerPixel, Saturation,
SceneCaptureType, SceneType, SensingMethod, Sharpness, ShutterSpeedValue,
Software, SpatialFrequencyResponse, SpectralSensitivity, StripByteCounts,
StripOffsets, SubfileType, SubjectArea, SubjectDistance,
SubjectDistanceRange, SubjectLocation, SubsecTime, SubsecTimeDigitized,
SubsecTimeOriginal, Threshholding, UserComment, WhiteBalance, XResolution,
YResolution.
}
\examples{
"ExposureTime" \%in\% supported_tags()
"AnInventedTag" \%in\% supported_tags()
}
\author{
Charles Martin
}
