% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getexif.R
\name{exifr}
\alias{exifr}
\title{Read EXIF data}
\usage{
exifr(filename, recursive = FALSE, quiet = TRUE, exiftoolargs = NULL)
}
\arguments{
\item{filename}{A vector of filenames (like that generated by
\code{list.files(..., full.names=TRUE)})}

\item{recursive}{If directory is specified by \code{filename}, process files recursively.}

\item{quiet}{\code{FALSE} if status updates are desired, \code{TRUE}
otherwise.}

\item{exiftoolargs}{a list of args to be passed to ExifTool (e.g. \code{c("-filename", "-imagesize")})}
}
\description{
This function has been deprecated, use \link{read_exif} instead.
}
