% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exif_old.R
\name{exiftool.call}
\alias{exiftool.call}
\title{Call exiftool from R}
\usage{
exiftool.call(args = c("--help"), fnames = NULL, intern = FALSE, ...)
}
\arguments{
\item{args}{a list of non-shell quoted arguments (e.g. \code{-n -csv})}

\item{fnames}{a list of filenames (\code{shQuote()} will be applied to this vector)}

\item{intern}{\code{TRUE} if output should be returned as a character vector.}

\item{...}{additional arguments to be passed to \code{system()}}
}
\description{
This function has been deprecated. Use \link{exiftool_call} instead.
}
