% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getexif.R
\name{exiftool.call}
\alias{exiftool.call}
\title{Call exiftool from R}
\usage{
exiftool.call(args = c("--help"), fnames = NULL, perlpath = NULL,
  intern = FALSE, ...)
}
\arguments{
\item{args}{a list of non-shell quoted arguments (e.g. \code{-n -csv})}

\item{fnames}{a list of filenames (\code{shQuote()} will be applied to this vector)}

\item{perlpath}{the location of the perl interpreter. \code{NULL} for autodetect.}

\item{intern}{\code{TRUE} if output should be returned as a character vector.}

\item{...}{additional arguments to be passed to \code{system()}}
}
\value{
The exit code if \code{intern=FALSE}, or the standard output as a character vector
 if \code{intern=TRUE}.
}
\description{
Uses \code{system()} to run a basic call to \code{exiftool}.
}
\examples{
exiftool.call("--help")
}

