% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{rational_to_numeric}
\alias{rational_to_numeric}
\title{Converts a rational number from to string to its numeric representation}
\usage{
rational_to_numeric(x)
}
\arguments{
\item{x}{A string containing the number to convert.}
}
\value{
The numerical value of x.
}
\description{
Converts a rational number from to string to its numeric representation
}
\examples{
rational_to_numeric("1/3200")
}

