% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exifr.R
\name{supported_tags}
\alias{supported_tags}
\title{List EXIF tags currently supported by this package.}
\usage{
supported_tags()
}
\value{
A vector of EXIF tag names.
}
\description{
List EXIF tags currently supported by this package.
}
\examples{
"ExposureTime" \%in\% supported_tags()
"AnInventedTag" \%in\% supported_tags()
}

